requirements.txt
discord.py
mss
pyautogui
pynput
requests
pyperclip
pillow
numpy
psutil